feat(event dispatcher): added a new field to the EventMethod.java class, EventMethod#requireCustomId which has the same effect as the old RequireCustomId.java annotation.
Also deprecated RequireCustomId.java in favour of the new field.

This change has been made because the old annotation had performance issues, as the annotation was checked at every event dispatch, and reflection takes time to run. Instead, the new field is checked when the event is registered to improve performance.